refactor: use dsp for eth
