+++ api/prove_mock.js
@@ -4,27 +4,24 @@ import { ZKGraphRequireFailed } from "../common/error.js";
 
 /**
  * Mock the zkwasm proving process for pre-test purpose.
- * @param {string} rpcUrl
- * @param {number | string} blockid
- * @param {string} expectedStateStr
- * @param {boolean} isLocal
- * @param {boolean} enableLog
+ * @param {Uint8Array} wasmUnit8Array 
+ * @param {string} privateInputStr 
+ * @param {string} publicInputStr 
  * @returns {boolean} - the mock testing result
  */
 export async function proveMock(wasmUnit8Array, privateInputStr, publicInputStr) {
-    // let [privateInputStr, publicInputStr] = await proveInputGen(yamlContent, rpcUrl, blockid, expectedStateStr, isLocal, enableLog)
     const mock = new ZKWASMMock();
     mock.set_private_input(privateInputStr);
     mock.set_public_input(publicInputStr);
     setupZKWasmMock(mock);
 
-    const { zkmain, runRegisterHandle } = await instantiateWasm(wasmUnit8Array).catch((error) => {
+    const { zkmain } = await instantiateWasm(wasmUnit8Array).catch((error) => {
         throw error
     });
 
     try {
-        runRegisterHandle()
-        zkmain();
+      // __as_start();
+      zkmain();
     } catch (e){
         if (e instanceof ZKGraphRequireFailed){
             return false
